Use `Cow<str>` to avoid unnecessary memory reallocations
